+++ test/libsolidity/syntaxTests/array/slice/calldata_dynamic_encode.sol
@@ -1,0 +1,7 @@
+contract C {
+    function f(uint256[] calldata x) external pure {
+        abi.encode(x[1:2]);
+    }
+}
+// ----
+// TypeError: (85-91): This type cannot be encoded.